fix: auto-seed reference data on deploy & add empty state to exercises
Fresh prod DBs had no exercises/tests/quotes/achievements because the
Dockerfile only ran migrations, not the seed. Split seed.ts into a
reusable seed-data module and a prod-safe seed-reference.ts that skips
demo-account creation, and wire it into the container CMD so reference
data is always present after migrations.

Also add an EmptyState fallback to ExercisesListPage so the page shows
a friendly message instead of a blank area if the list is ever empty.
